% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{BIGverse_deps}
\alias{BIGverse_deps}
\title{List all BIGverse dependencies}
\usage{
BIGverse_deps(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
BIGverse packages.}
}
\description{
List all BIGverse dependencies
}
