% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{BIGverse_packages}
\alias{BIGverse_packages}
\title{List all packages in the BIGverse}
\usage{
BIGverse_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include BIGverse in the list?}
}
\description{
List all packages in the BIGverse
}
\examples{
BIGverse_packages()
}
