% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verse.R
\name{BIGverse_update}
\alias{BIGverse_update}
\title{Update BIGverse packages}
\usage{
BIGverse_update(recursive = FALSE)
}
\arguments{
\item{recursive}{If \code{TRUE}, will also check all dependencies of
BIGverse packages.}
}
\description{
This will check to see if all BIGverse packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
BIGverse_update()
}
}
